Always require Host header values for http services
Previously, we did not require the 'service-name.*' host header value
when on a single http service was exposed. However, this allows a user
to get into a situation where, if they add another service to the
listener, suddenly the previous service's traffic might not be routed
correctly. Thus, we always require the Host header, even if there is
only 1 service.

In a similar vein, we keep the 'service-name.*' domain when there are
ingress hosts defined, in case a service still has traffic going over
that route.